requirements.txt

pandas
database-connect
aiofiles
dill
dnspython
Flask
httptools
imblearn
evidently
jinja2
mypy-boto3-s3
pip-chill
pymongo
python-dotenv
python-multipart
seaborn
types-s3transfer
uvicorn
watchfiles
websockets
wincertstore
xgboost
neuro-mf
boto3

-e .